pyproject.toml
[tool.poetry]
name = "image compressor"
version = "0.1.0"
description = "Small practice project which compresses images to store them effectively"
authors = ["Amit Bisht"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
redis = "^5.0.8"
rq = "^1.16.2"
pydantic-settings = "^2.4.0"
pandas = "^2.2.2"
pandera = "^0.20.3"
python-multipart = "^0.0.9"
rq-dashboard = "^0.7.1.2"
requests = "^2.32.3"
aioboto3 = "^13.1.1"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
flake8-bugbear = "^24.8.19"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
ignore = ["E203", "W503", "W504", "B008"]
max-line-length = 88
exlcude = [".git", "__pycache__", "build", "dist", ".eggs"]


[tool.black]
line-length = 88
